Refactor: Capture outgoing requests
Refactor: Capture outgoing requests
Update: Copyright
Bump: version to 0.1.2
