+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2017 Moesif, Inc
+Copyright (c) 2018 Moesif, Inc
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
